Move serialize function to utils
