+++ src/quoalise/utils.py
@@ -1,4 +1,5 @@
 import datetime as dt
+from .data import Record
 
 
 def parse_iso_date(date_str):
@@ -7,3 +8,11 @@ def parse_iso_date(date_str):
 
 def format_iso_date(date):
     return date.strftime("%Y-%m-%d")
+
+
+def serialize(obj):
+    if isinstance(obj, (dt.datetime, dt.date)):
+        return obj.isoformat()
+    if isinstance(obj, Record):
+        return obj.__dict__
+    return str(obj)
